Moving functions to utils files
